fmt: Rename case of clockify_api and toggl_api
